requirements.txt
Flask>=3.0,<4.0
pymongo>=4.6,<5.0
websockets>=12.0,<14.0
aiohttp>=3.9,<4.0
protobuf>=4.25,<6.0
python-dotenv>=1.0,<2.0
